+++ modules/index.all.esm.mjs
@@ -3,6 +3,7 @@
  */
 export * from './head/index.esm.mjs';
 export * from './base/index.esm.mjs';
+export * from './controllers/index.esm.mjs';
 export * from './templating/index.esm.mjs';
 export * from './components/index.esm.mjs';
 export * from './extra/index.esm.mjs';