+++ modules/index.all.mjs
@@ -3,6 +3,7 @@
  */
 import './head/index.mjs';
 import './base/index.mjs';
+import './controllers/index.mjs';
 import './templating/index.mjs';
 import './components/index.mjs';
 import './extra/index.mjs';