+++ modules/index.esm.mjs
@@ -3,4 +3,5 @@
  */
 export * from './head/index.esm.mjs';
 export * from './base/index.esm.mjs';
+export * from './controllers/index.esm.mjs';
 export * from './templating/index.esm.mjs';