+++ modules/index.mjs
@@ -3,4 +3,5 @@
  */
 import './head/index.mjs';
 import './base/index.mjs';
+import './controllers/index.mjs';
 import './templating/index.mjs';