+++ CHANGES.md
@@ -10,6 +10,8 @@
  - Added - GSTheme added theme getter
  - Added - GSDynamicStyle - dynamic CSS style
  - Added - GSAttr generic element attribute proxy 
+ - Added - Controller support for easy component extensions
+ - Added - Scheduled attribute change updates
  - Improvement - GSEvents registration
 
 # Release v1.9.0 (04.12.2023.)